Fixed call instruction alignment, changed generation to ret mapping
When testing the driver, I changed the base address and offset of
the code to be executed.  This caused an incorrect offset to be
executed at first, but it showed that call instructions were not
being generated correctly, and that their offset was not correctly
aligned on chunk boundaries.  I fixed this.

Also, I needed the mapping to determine what offset to call in the
driver, so I changed the code generation to return the mapping
instead of the code.  Since the code is not malloc'd anymore, there
wasn't really a reason to return it.

+++ miniverse.h
@@ -1,4 +1,4 @@
 #include <ssdis.h>
 
-uint8_t* gen_code(const uint8_t* bytes, size_t bytes_size, uintptr_t address, uintptr_t new_address,
+uint32_t* gen_code(const uint8_t* bytes, size_t bytes_size, uintptr_t address, uintptr_t new_address,
     size_t *new_size, uint8_t chunk_size, bool (*is_target)(uintptr_t address, uint8_t *bytes));
